server/requirements.txt
dotenv
fastapi
google-genai
pydantic
uvicorn
